@@ -1,6 +1,9 @@
-package com.example.ledgr
+package com.example.ledgr.ui.widget.date
 
 import android.util.Log
+import java.time.Instant
+import java.time.ZoneId
+import java.time.ZonedDateTime
 import java.util.*
 
 class Date() {
@@ -8,14 +11,16 @@ class Date() {
     val year = currentInstant.get(Calendar.YEAR)
     val day = currentInstant.get(Calendar.DAY_OF_MONTH)
     val month = currentInstant.get(Calendar.MONTH)
-    val monthDisplayName = currentInstant.getDisplayName(Calendar.MONTH, Calendar.LONG, Locale.ROOT).toString()
+    private val monthDisplayName = currentInstant.getDisplayName(Calendar.MONTH, Calendar.LONG, Locale.ROOT)!!.toString()
+    private val zdt: ZonedDateTime =
+        ZonedDateTime.ofInstant(Instant.now(), ZoneId.of("America/Montreal"))
 
     fun getCompleteDate(date: String):String {
-        Log.d("acaligetCompleteDate:", date.toString())
+        Log.d("acaligetCompleteDate:", date)
         val months = currentInstant.getDisplayNames(Calendar.MONTH, Calendar.LONG, Locale.ROOT)
         val dateArray = date.split(' ')
         Log.d("acaligetCompleteDate:", dateArray.toString())
-        val monthAsInt: Int? = months[dateArray[1]]
+        val monthAsInt: Int? = months!![dateArray[1]]
 
         if (monthAsInt != null) {
             return if (monthAsInt+1 >= 10) {
@@ -37,6 +42,15 @@ class Date() {
         val dateArray = date.split('-')
         val monthString = months[dateArray[1].toInt() -1]
         return "${dateArray[2]} $monthString ${dateArray[0]}"
+    }
+
+    fun getCurrentMonth(): String {
+
+        return zdt.month.toString().toLowerCase(Locale.ROOT).capitalize(Locale.ROOT)
+    }
+
+    fun getCurrentYear(): String {
+        return zdt.year.toString()
     }
 
 
