@@ -1,17 +1,22 @@
 package com.example.ledgr
 
+import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.util.Log
-import android.widget.ArrayAdapter
 import android.widget.ListView
 import android.widget.TextView
 import java.util.*
-import com.example.ledgr.data.LedgrDataSource
+import com.example.ledgr.ui.widget.date.Date
+
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModelProvider
+import com.example.ledgr.data.model.Transaction
+import com.example.ledgr.ui.transactions.TransactionsViewModel
+import com.example.ledgr.ui.transactions.TransactionsViewModelFactory
 import com.google.gson.JsonArray
 import kotlinx.android.synthetic.main.fragment_view_transactions.*
 
@@ -23,8 +28,9 @@ import kotlinx.android.synthetic.main.fragment_view_transactions.*
 class ViewTransactionsFragment : Fragment() {
     // private val transactions = ArrayList<String>()
     var transactions: MutableLiveData<Any> = MutableLiveData()
-    private val transactionList = ArrayList<Map<*, *>>()
-    private lateinit var transactionListAdapter: ArrayAdapter<String>
+    private val transactionList = ArrayList<Transaction>()
+
+    private lateinit var transactionsViewModel: TransactionsViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -36,17 +42,35 @@ class ViewTransactionsFragment : Fragment() {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        Log.d("acalibundle", arguments?.getString("api").toString().removeSurrounding("\""))
-        val apiKey = arguments?.getString("api").toString().removeSurrounding("\"")
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        // Log.d("acalibundle", arguments?.getString("api").toString().removeSurrounding("\""))
+        // val token = arguments?.getString("api").toString().removeSurrounding("\"")
         // ListView
         val transactionsList = activity?.findViewById<ListView>(R.id.transaction_list)
 
+        val sharedPref = activity?.getSharedPreferences(getString(R.string.api_token), Context.MODE_PRIVATE) ?: return
+        val token = sharedPref.getString(getString(R.string.api_token), "")
 
+
+        val url = "https://api.ledgr.site/transactions?month=${Date().getCurrentMonth()}&year=${Date().getCurrentYear()}"
+
+        // Initialize Transactions ViewModel
+        transactionsViewModel = ViewModelProvider(this, TransactionsViewModelFactory(requireActivity(), token!!.toString())).get(TransactionsViewModel::class.java)
+
+        // Connect to Ledgr Database
+        transactionsViewModel.get(url)
+
+        /** Depreciated
         val ledgr =
-            LedgrDataSource(apiKey, requireActivity())
+        LedgrDataSource(requireActivity(), token)
         ledgr.transactions().getLegacy(transactions)
+         */
 
-        transactions.observe(viewLifecycleOwner, androidx.lifecycle.Observer {
+        transactionsViewModel.transactions.observe(viewLifecycleOwner, androidx.lifecycle.Observer {
 
             Log.d("acaliObserveTransList", it.toString())
             val check = it as JsonArray
@@ -59,17 +83,35 @@ class ViewTransactionsFragment : Fragment() {
                         "${item.asJsonObject.get("amount")}.00"
                     ).count() == 1
                 ) "$${item.asJsonObject.get("amount").asString}0" else "$${item.asJsonObject.get("amount").asString}"
+                /**
                 val tMap = mapOf<String, String>(
-                    "date" to item.asJsonObject.get("date").asString,
-                    "category" to item.asJsonObject.get("category").asString,
-                    "description" to item.asJsonObject.get("description").asString,
-                    "amount" to amount
+                "date" to item.asJsonObject.get("date").asString,
+                "category" to item.asJsonObject.get("category").asString,
+                "description" to item.asJsonObject.get("description").asString,
+                "amount" to amount
                 )
 
-                transactionList.add(tMap)
+                val transaction = Transaction(
+                date=item.asJsonObject.get("date").asString,
+                category=item.asJsonObject.get("category").asString,
+                description = item.asJsonObject.get("description").asString,
+                amount = item.asJsonObject.get("amount").asFloat
+                )
 
+                transactionList.add(transaction)
+                 */
+
+                transactionList.add(Transaction(
+                    date=Date().displayAsString(item.asJsonObject.get("date").asString),
+                    category=item.asJsonObject.get("category").asString,
+                    description = item.asJsonObject.get("description").asString,
+                    amount = item.asJsonObject.get("amount").asFloat
+                ))
+                val dateTest = Date().displayAsString(item.asJsonObject.get("date").asString)
+                Log.d("acaliDatetest", dateTest)
             }
-            Log.d("acali.arrayMap", transactionsList.toString())
+
+            Log.i("acali.arrayMap", transactionsList.toString())
 
             val transactionListAdapterClass =
                 TransactionListAdapter(requireActivity(), transactionList)
@@ -82,10 +124,36 @@ class ViewTransactionsFragment : Fragment() {
         transaction_list_refresh.setOnRefreshListener {
             Log.d("acaliONREFRESH", "onRefresh called from SwipeRefreshLayout")
             transactionList.clear()
-            ledgr.transactions().getLegacy(transactions)
+            // ledgr.transactions().getLegacy(transactions)
+            transactionsViewModel.get(url)
             transaction_list_refresh.isRefreshing = false
         }
     }
 
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        Log.d("acali-ViewTransactionsFragment", "onCreate was called")
+    }
+
+    override fun onPause() {
+        super.onPause()
+        Log.d("acali-ViewTransactionsFragment", "onPause was called")
+    }
+
+    override fun onStop() {
+        super.onStop()
+        Log.d("acali-ViewTransactionsFragment", "onStop was called")
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        Log.d("acali-ViewTransactionsFragment", "onResume was called")
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        Log.d("acali-ViewTransactionsFragment", "onDestroy was called")
+    }
 
 }
