@@ -2,18 +2,25 @@ package com.example.ledgr.ui.budget
 
 import android.app.Activity
 import android.content.Context
+import android.content.Intent
+import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
-import android.widget.ProgressBar
+import android.widget.ImageView
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.res.ResourcesCompat
+import androidx.fragment.app.Fragment
 import com.example.ledgr.R
 import com.example.ledgr.data.model.BudgetCategory
-import kotlin.math.roundToInt
 
-class BudgetProgressListAdapter(private val context: Activity, private val budget: ArrayList<BudgetCategory>)
+class BudgetProgressListAdapter(
+    private val context: Activity,
+    private val budget: ArrayList<BudgetCategory>
+)
     : BaseAdapter() {
     override fun getCount(): Int {
         return budget.size
@@ -29,11 +36,16 @@ class BudgetProgressListAdapter(private val context: Activity, private val budge
 
     override fun getView(position: Int, view: View?, parent: ViewGroup): View {
         val inflator = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
-        val rowView = inflator.inflate(R.layout.budget_row, parent, false)
+        val rowView = inflator.inflate(R.layout.budget_item, parent, false)
 
-        val progressBar = rowView.findViewById(R.id.progress_bar) as ProgressBar
+        // val progressBar = rowView.findViewById(R.id.progress_bar) as ProgressBar
         val categoryText = rowView.findViewById(R.id.budget_category) as TextView
         val plannedText = rowView.findViewById(R.id.budget_planned) as TextView
+        val icon = rowView.findViewById(R.id.budget_icon) as ImageView
+        // val progressBarTest = rowView.findViewById(R.id.progress_bar_example) as com.example.ledgr.ui.widget.ProgressBar
+
+
+
 
         /**
          * Depreciated Version
@@ -48,15 +60,61 @@ class BudgetProgressListAdapter(private val context: Activity, private val budge
 
         val budget : BudgetCategory = getItem(position) as BudgetCategory
 
-        val label = "$${budget.actual} of $${budget.planned}"
+        val label = "$${budget.formatActualForDisplay()} of $${budget.formatPlannedForDisplay()}"
         categoryText.text = budget.category
         plannedText.text = label
-        progressBar.max = budget.planned.roundToInt()
-        progressBar.progress = budget.actual.toString().toFloat().roundToInt()
+        val draw = context.resources.getIdentifier(budget.icon, "drawable", context.packageName)
+
+        // progressBar.max = budget.planned.roundToInt()
+        // progressBar.progress = budget.actual.toString().toFloat().roundToInt()
+        val budgetIcon = "restaurant"
+        val iconMap = mapOf<String, Int>(
+            "restaurant" to R.drawable.restaurant,
+            "build" to R.drawable.build
+        )
+        /**
+        icon.setImageDrawable(
+            iconMap[budgetIcon]?.let {
+                ResourcesCompat.getDrawable(
+                    context.resources,
+                    it,
+                    null)
+            })
+        */
+        icon.setImageDrawable(ResourcesCompat.getDrawable(context.resources, draw, null))
+        /**
+        progressBarTest.apply {
+            setActual(budget.actual)
+            setPlanned(budget.planned)
+        }
+        */
+
+
 
         rowView.setOnClickListener {
             val budgetToastText = "${categoryText.text}: $label"
             Log.i("acali-rowView", budgetToastText)
+
+            val bundle = Bundle().apply {
+                this.putString("category", categoryText.text.toString())
+            }
+
+            val targetFragment : Fragment = BudgetDetailFragment::class.java.newInstance()
+            targetFragment.arguments = bundle
+
+            val transaction = context as AppCompatActivity
+            transaction.supportFragmentManager.beginTransaction().replace(R.id.layout_frame, targetFragment).addToBackStack(categoryText.text.toString()).commit()
+
+
+            //transaction.replace(R.id.layout_frame, targetFragment).commit()
+            /**
+            val myIntent = Intent(context, BudgetDetailsActivity::class.java)
+
+
+            myIntent.putExtra("category", categoryText.text.toString())
+
+            context.startActivity(myIntent)
+            */
         }
 
         return rowView
