@@ -1,28 +1,31 @@
 package com.example.ledgr.ui.dashboard
 
+import android.animation.ObjectAnimator
 import android.annotation.SuppressLint
+import android.content.Context
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.view.animation.BounceInterpolator
+import android.view.animation.DecelerateInterpolator
 import android.widget.ArrayAdapter
-import android.widget.ListView
-import android.widget.TextView
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.example.ledgr.R
-import com.example.ledgr.data.LedgrRepository
-import com.example.ledgr.data.model.BudgetCategory
+import com.example.ledgr.data.DataBuilder
+import com.example.ledgr.ui.widget.date.Date
 import com.example.ledgr.ui.budget.BudgetProgressListAdapter
+import com.google.gson.JsonArray
 import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.budget_row.view.*
+import kotlinx.android.synthetic.main.budget_item.view.*
 import kotlinx.android.synthetic.main.fragment_dashboard.*
 import kotlinx.android.synthetic.main.fragment_new_transaction.*
+
 import java.util.*
-import kotlin.collections.ArrayList
+import kotlin.math.round
 
 /**
  * A simple [Fragment] subclass.
@@ -31,13 +34,14 @@ import kotlin.collections.ArrayList
  */
 class DashboardFragment : Fragment() {
     // var spending: MutableLiveData<Any> = MutableLiveData()
-    var budget: MutableLiveData<Any> = MutableLiveData()
+    // var budget: MutableLiveData<Any> = MutableLiveData()
 
     // private val budgetList = ArrayList<Map<String, String>>()
-    private lateinit var budgetListAdapter : ArrayAdapter<String>
+    // private lateinit var budgetListAdapter: ArrayAdapter<String>
 
-    private lateinit var dashboardViewModel : DashboardViewModel
-    private lateinit var dashboardViewModelFactory: DashboardViewModelFactory
+    private lateinit var dashboardViewModel: DashboardViewModel
+    private lateinit var dataRepository: DataBuilder
+    //private lateinit var dashboardViewModelFactory: DashboardViewModelFactory
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -47,81 +51,127 @@ class DashboardFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_dashboard, container, false)
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        // Spending by categories ListView
-        val budgetlist = activity?.findViewById<ListView>(R.id.budget_list)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        Log.d("acaliDASHBOARD_FRAGMENT", "onViewCreated was called")
 
-        dashboardViewModelFactory = DashboardViewModelFactory()
-        dashboardViewModel = ViewModelProvider(this, dashboardViewModelFactory).get(DashboardViewModel::class.java)
-
-        dashboardViewModel.getData(LedgrRepository("LHWmiGNoaVbrgYTv7qETIVpoNkJ8H9IB1Y3Ze72voXY5Oei8Pyl7gp2Apfpw", requireActivity()))
-
-        dashboardViewModel.budget.observe(viewLifecycleOwner, {
-            /**
-             * Depreciated
-             *
-            for (item in it as JsonArray) {
-
-                // val outputLine = "${item.asJsonObject.get("category").asString} $${item.asJsonObject.get("planned").asString}"
-                val budget = mapOf<String, String>(
-                    "category" to item.asJsonObject.get("category").asString,
-                    "planned" to item.asJsonObject.get("planned").asString,
-                    "actual" to item.asJsonObject.get("actual").asString
-                )
-                budgetList.add(budget)
-            }
-            **/
-            // budgetListAdapter = ArrayAdapter(requireActivity(), android.R.layout.simple_list_item_1, budgetList)
-            val budgetListAdapter = BudgetProgressListAdapter(requireActivity(), it as ArrayList<BudgetCategory>)
-            budgetlist?.adapter = budgetListAdapter
-        })
-
-        val spendingDisplay = activity?.findViewById<TextView>(R.id.spending_display)
-
-        dashboardViewModel.spending.observe(viewLifecycleOwner, Observer {
-            val display = "$${it}"
-            spendingDisplay?.text = display
-            // dashboardViewModel.spending.value = it.toString().toDouble()
-            // spendingDisplay?.text = "$${dashboardViewModel.spending.toString()}"
-            // progress_bar.progress = it.toString().substringAfter("$").toFloat().roundToInt()
-        })
     }
 
     @SuppressLint("SetTextI18n")
     override fun onResume() {
         super.onResume()
 
-        val spendingDisplay = activity?.findViewById<TextView>(R.id.spending_display)
-        val budgetlist = activity?.findViewById<ListView>(R.id.budget_list)
+        Log.d("acaliDASHBOARD_FRAGMENT:", "onResume was called")
+
+        // Spending by categories ListView
+        // val budgetlist = activity?.findViewById<ListView>(R.id.budget_list)
+        // val spendingDisplay = activity?.spending_display
+        // val token = arguments?.getString("api").toString()
+        // Get logged in user's api token.
+        val sharedPref =
+            activity?.getSharedPreferences(getString(R.string.api_token), Context.MODE_PRIVATE)
+                ?: return
+        val token = sharedPref.getString(getString(R.string.api_token), "")
+        val url =
+            "https://api.ledgr.site/budget?month=${Date().getCurrentMonth()}&year=${Date().getCurrentYear()}"
+
+
+        dataRepository = DataBuilder()
+
+        // Initialize Dashboard ViewModel
+        dashboardViewModel =
+            ViewModelProvider(this, DashboardViewModelFactory(requireActivity(), token!!.toString())).get(
+                DashboardViewModel::class.java
+            )
+
+        // Connect to Ledgr Database
+        dashboardViewModel.get(url)
 
         dashboardViewModel.budget.observe(viewLifecycleOwner, Observer {
+            val budgetList: JsonArray = it as JsonArray
+            val viewList = dataRepository.budgetListForView(budgetList)
+
             /**
-             * Depreciated
-            for (item in it as JsonArray) {
-                // val outputLine = "${item.asJsonObject.get("category").asString} $${item.asJsonObject.get("planned").asString}"
-                val budget = mapOf<String, String>(
-                        "category" to item.asJsonObject.get("category").asString,
-                        "planned" to item.asJsonObject.get("planned").asString,
-                        "actual" to item.asJsonObject.get("actual").asString
-                )
-                // budgetList.add(budget)
+            val viewList: ArrayList<BudgetCategory> = ArrayList()
+
+            budgetList.forEach {
+            val budgetItem = it.asJsonObject
+            if (budgetItem.get("planned").asFloat > 0F) {
+            viewList.add(
+            BudgetCategory(
+            budgetItem.get("category").asString,
+            budgetItem.get("planned").asFloat,
+            budgetItem.get("actual").asFloat,
+            budgetItem.get("icon").asString,
+            )
+            )
+            }
+            }
+            /**
+
+            for (budgetItem in budgetList) {
+            val budget = BudgetCategory(
+            budgetItem.asJsonObject.get("category").asString,
+            budgetItem.asJsonObject.get("planned").asFloat,
+            budgetItem.asJsonObject.get("actual").asFloat
+            )
+            Log.i("acali-dashboardViewModel.budget.observe", budget.toString())
+            viewList.add(budget)
+            val key = budgetItem.asJsonObject.get("period").asString
+            if (budgetPeriodSet.containsKey(key)) {
+            val tempValue = budgetPeriodSet.get(key)
+            if (tempValue != null) {
+            budgetPeriodSet[key] =
+            tempValue + budgetItem.asJsonObject.get("actual").asFloat
+            }
+            } else {
+            budgetPeriodSet[key] = budgetItem.asJsonObject.get("actual").asFloat
+            }
+
             }
             */
-            // budgetListAdapter = ArrayAdapter(requireActivity(), android.R.layout.simple_list_item_1, budgetList)
-            val budgetListAdapter = BudgetProgressListAdapter(requireActivity(), it as ArrayList<BudgetCategory>)
-            budgetlist?.adapter = budgetListAdapter
+             */
+            val budgetListAdapter = BudgetProgressListAdapter(
+                requireActivity(),
+                viewList
+            )
+            budget_list.adapter = budgetListAdapter
+            // budgetlist?.adapter = budgetListAdapter
         })
 
         dashboardViewModel.spending.observe(viewLifecycleOwner, Observer {
-            val display = "$${it}"
-            spendingDisplay?.text = display
-            // dashboardViewModel.spending.value = it.toString().toDouble()
-            // spendingDisplay?.text = "$${dashboardViewModel.spending.toString()}"
-            // progress_bar.progress = it.toString().substringAfter("$").toFloat().roundToInt()
+            var totalSpending = 0F
+            var plannedSpending = 0F
+            val budgetList: JsonArray = it as JsonArray
+            for (budgetItem in budgetList) {
+                totalSpending += budgetItem.asJsonObject.get("actual").asFloat
+                plannedSpending += budgetItem.asJsonObject.get("planned").asFloat
+            }
+            val roundedTotal = round(totalSpending)
+            progress_bar_test.apply {
+                setActual(0F)
+                setPlanned(plannedSpending)
+            }
+            ObjectAnimator.ofFloat(progress_bar_test, "actual", 0F, totalSpending.toFloat()).apply {
+                duration = 2000
+                interpolator = DecelerateInterpolator()
+                start()
+            }
+            // progress_bar_test
+            val display = "$${roundedTotal}"
+            // spendingDisplay?.setFormattedText(display)
+            spending_display.setFormattedText(display)
         })
+    }
 
-        Log.d("acaliDASHBOARD_FRAGMENT:", "onResume was called")
+    override fun onStart() {
+        super.onStart()
+        Log.d("acaliDASHBOARD_FRAGMENT", "onStart was called")
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        Log.d("acaliDASHBOARD_FRAGMENT", "onCreate was called")
     }
 
     override fun onPause() {
@@ -136,11 +186,18 @@ class DashboardFragment : Fragment() {
 
     override fun onDestroyView() {
         super.onDestroyView()
-        Log.i("acaliDASHBOARD_FRAGMENT", "onDestroyView called")
+        Log.d("acaliDASHBOARD_FRAGMENT", "onDestroyView called")
     }
+
     override fun onDetach() {
         super.onDetach()
-        Log.i("acaliDASHBOARD_FRAGMENT", "onDetach called")
+        Log.d("acaliDASHBOARD_FRAGMENT", "onDetach called")
+    }
+
+    override fun onViewStateRestored(savedInstanceState: Bundle?) {
+        super.onViewStateRestored(savedInstanceState)
+        Log.d("acaliDASHBOARD_FRAGMENT", "onViewStateRestored was called")
+
     }
 
 }
